Cleanup migrations, clean commands to create new DB from old DB
